move compile_fail tests to diman_unit_system where they belong
