+++ Cargo.toml
@@ -49,5 +49,3 @@ diman_unit_system = { path = "crates/diman_unit_system", version = "0.4" }
 [dev-dependencies]
 serde_yaml = "0.9.27"
 lazy_static = "1.4.0"
-trybuild = "1.0"
-
