+++ crates/diman_unit_system/Cargo.toml
@@ -33,3 +33,8 @@ proc-macro2 = "1.0"
 
 [lib]
 proc-macro = true
+
+[dev-dependencies]
+trybuild = "1.0"
+num-traits = { version = "0.2.17", default-features = false }
+glam = { version = "0.24.2"}
