+++ crates/diman_unit_system/tests/compile_fail/dimension_annotation_violated_constant.stderr
