+++ crates/diman_unit_system/tests/compile_fail/dimension_annotation_violated_unit.rs
@@ -1,15 +1,16 @@
 #![allow(incomplete_features)]
 #![feature(generic_const_exprs, adt_const_params)]
 
-use ::diman::unit_system;
+use diman_unit_system::unit_system;
 unit_system!(
     quantity_type Quantity;
     dimension_type Dimension;
     dimension Mass;
+    dimension Length;
     #[base(Mass)]
     #[symbol(kg)]
-    unit kilograms;
-    unit foo: kilograms = kilograms;
+    unit kilograms: Mass;
+    unit foo: Length = kilograms * kilograms;
 );
 
 fn main() {
