+++ crates/diman_unit_system/tests/compile_fail/dimension_annotation_with_unit_not_allowed.rs
@@ -1,12 +1,15 @@
 #![allow(incomplete_features)]
 #![feature(generic_const_exprs, adt_const_params)]
 
-use ::diman::unit_system;
+use diman_unit_system::unit_system;
 unit_system!(
     quantity_type Quantity;
     dimension_type Dimension;
     dimension Mass;
-    constant SOLAR_MASS: Mass = 1.988477e30 * Mass;
+    #[base(Mass)]
+    #[symbol(kg)]
+    unit kilograms;
+    unit foo: kilograms = kilograms;
 );
 
 fn main() {
