+++ crates/diman_unit_system/tests/compile_fail/dimension_annotation_with_unit_not_allowed.stderr
