+++ crates/diman_unit_system/tests/compile_fail/dimension_definition_with_numeric_factor.rs
@@ -1,13 +1,14 @@
 #![allow(incomplete_features)]
 #![feature(generic_const_exprs, adt_const_params)]
 
-use diman::unit_system;
+use diman_unit_system::unit_system;
 
 unit_system!(
     quantity_type Quantity;
     dimension_type Dimension;
-    #[prefix(kilo)]
-    unit foo = undefined;
+    dimension Dimensionless = 1;
+    dimension Length = 2.0 / Dimensionless;
 );
 
-fn main() {}
+fn main() {
+}
