+++ crates/diman_unit_system/tests/compile_fail/dimension_definition_with_numeric_factor.stderr
