+++ crates/diman_unit_system/tests/compile_fail/example_system/mod.rs
@@ -1,4 +1,4 @@
-use ::diman::unit_system;
+use diman_unit_system::unit_system;
 
 
 unit_system!(
