+++ crates/diman_unit_system/tests/compile_fail/float_unknown_method_on_quantity_type_alias.stderr
