+++ crates/diman_unit_system/tests/compile_fail/resolver_dimension_in_constant_expression_not_allowed.rs
@@ -1,16 +1,12 @@
 #![allow(incomplete_features)]
 #![feature(generic_const_exprs, adt_const_params)]
 
-use ::diman::unit_system;
+use diman_unit_system::unit_system;
 unit_system!(
     quantity_type Quantity;
     dimension_type Dimension;
     dimension Mass;
-    dimension Length;
-    #[base(Mass)]
-    #[symbol(kg)]
-    unit kilograms: Mass;
-    unit foo: Length = kilograms * kilograms;
+    constant SOLAR_MASS: Mass = 1.988477e30 * Mass;
 );
 
 fn main() {
