+++ crates/diman_unit_system/tests/compile_fail/resolver_dimension_in_unit_expression_not_allowed.rs
@@ -1,13 +1,12 @@
 #![allow(incomplete_features)]
 #![feature(generic_const_exprs, adt_const_params)]
 
-use ::diman::unit_system;
-
+use diman_unit_system::unit_system;
 unit_system!(
     quantity_type Quantity;
     dimension_type Dimension;
-    dimension Dimensionless = 1;
-    dimension Length = 2.0 / Dimensionless;
+    dimension Mass;
+    unit kilograms = 5.0 * Mass;
 );
 
 fn main() {
