+++ crates/diman_unit_system/tests/compile_fail/resolver_duplicate_base_unit.rs
@@ -1,7 +1,7 @@
 #![allow(incomplete_features)]
 #![feature(generic_const_exprs, adt_const_params)]
 
-use diman::unit_system;
+use diman_unit_system::unit_system;
 
 unit_system!(
     quantity_type Quantity;
@@ -10,8 +10,9 @@ unit_system!(
     #[base(Length)]
     #[symbol(m)]
     unit meters: Length;
-    unit foo: Length = 10.0 * meters;
-    unit foo: Length = 20.0 * meters;
+    #[base(Length)]
+    #[symbol(pc)]
+    unit parsec: Length;
 );
 
 fn main() {}
