+++ crates/diman_unit_system/tests/compile_fail/resolver_prefixed_or_aliased_units_dont_generate_multiple_errors.rs
@@ -1,13 +1,13 @@
 #![allow(incomplete_features)]
 #![feature(generic_const_exprs, adt_const_params)]
 
-use diman::unit_system;
+use diman_unit_system::unit_system;
 
 unit_system!(
     quantity_type Quantity;
     dimension_type Dimension;
-    dimension Length;
-    unit kilometers = 1000.0 * meters;
+    #[prefix(kilo)]
+    unit foo = undefined;
 );
 
 fn main() {}
