+++ crates/diman_unit_system/tests/compile_fail/resolver_unresolvable.rs
@@ -1,13 +1,14 @@
 #![allow(incomplete_features)]
 #![feature(generic_const_exprs, adt_const_params)]
 
-use ::diman::unit_system;
+use diman_unit_system::unit_system;
+
 unit_system!(
     quantity_type Quantity;
     dimension_type Dimension;
-    dimension Mass;
-    unit kilograms = 5.0 * Mass;
+    dimension Length;
+    unit meters = 0.001 * kilometers;
+    unit kilometers = 1000.0 * meters;
 );
 
-fn main() {
-}
+fn main() {}
