--- tests/compile_fail/glam_unknown_method_on_quantity_type_alias.rs
@@ -1,8 +1,0 @@
-#![allow(incomplete_features)]
-#![feature(generic_const_exprs, adt_const_params)]
-pub mod example_system;
-
-fn main() {
-    use example_system::vec2::Length;
-    Length::unknown_method(49.0);
-}
