--- tests/compile_fail/glam_unknown_method_on_quantity_type_alias.stderr
@@ -1,21 +1,0 @@
-error[E0599]: no function or associated item named `unknown_method` found for struct `example_system::Quantity` in the current scope
- --> tests/compile_fail/glam_unknown_method_on_quantity_type_alias.rs:7:13
-  |
-7 |     Length::unknown_method(49.0);
-  |             ^^^^^^^^^^^^^^ function or associated item not found in `Quantity<Vec2, Dimension>`
-  |
- ::: tests/compile_fail/example_system/mod.rs
-  |
-  |     quantity_type Quantity;
-  |                   -------- function or associated item `unknown_method` not found for this struct
-  |
-note: if you're trying to build a new `example_system::Quantity<Vec2, Dimension>` consider using one of the following associated functions:
-      example_system::Quantity::<Vec2, Dimension>::dimensionless
-      example_system::Quantity::<Vec2, Dimension>::grams
-      example_system::Quantity::<Vec2, Dimension>::joules
-      example_system::Quantity::<Vec2, Dimension>::kilograms
-      and $N others
- --> tests/compile_fail/example_system/mod.rs
-  |
-  |     dimension_type Dimension;
-  |                    ^^^^^^^^^
